chore: Custom row component to handle different file inputs
